refactor NotesSection
