+++ src/components/Navbar/NotesSection/NotesSection.module.scss
@@ -1,20 +1,17 @@
+@use "./../../../abstracts/mixins" as *;
+
 .mainBody {
-  display: flex;
   background-color: #fff;
-  padding: 1.5625rem 0;
   border-bottom: 2px solid #eee;
+  line-height: 1.5;
   margin: 0 2rem;
+  overflow-y: auto;
+  padding: 1.5625rem 0;
 
   p {
     margin-block-start: 0;
     margin-block-end: 0;
   }
-}
-
-.mainBodyContent {
-  height: 100%;
-  line-height: 1.5;
-  overflow-y: auto;
 }
 
 .mainBodyTitle {
@@ -24,25 +21,13 @@
 }
 
 .mainBodyTextWrapper {
-  display: flex;
-  flex-direction: column;
-  margin-top: 1.5rem;
-
-  .smallUp & {
-    flex-direction: row;
-    gap: 5%;
-  }
-}
-
-.mainBodyText {
-  //max-width: 40em;
-  //width: 80%;
+  margin-top: 1rem;
 }
 
 .mainBodyButtons {
   display: flex;
   flex-direction: column;
-  gap: 1rem 1.5rem;
+  gap: 1rem;
   margin-top: 1.5rem;
   width: 100%;
 
@@ -62,4 +47,49 @@
   &:hover {
     background-color: var(--theme-color-3);
   }
+}
+
+.sectionTitle {
+  background-color: transparent;
+  border: none;
+  border-radius: 2rem;
+  cursor: pointer;
+  color: #ffffff;
+  padding: 0.5rem 1rem;
+  white-space: nowrap;
+
+  &:hover {
+    background-color: var(--theme-color-3);
+  }
+
+  &.active {
+    text-decoration: underline;
+  }
+}
+
+.notesList {
+  @media print {
+    max-height: unset !important;
+  }
+}
+
+.overlay {
+  @include modal-overlay;
+}
+
+.dialogContent {
+  @include modal-content;
+  padding: 0;
+  overflow: unset;
+}
+
+.closeButton {
+  @include modal-close-button;
+}
+
+.contentWrapper {
+  flex: 1;
+  height: 100%;
+  width: 100%;
+  overflow-y: auto;
 }
